dia4/web_asistencia/requirements.txt
mysql-connector-python
flask
pip install flask-mysqldb